+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - ParallelExecution adds a HttpException instead of an empty Result in case of an endpoint failure
+- Facet pivot stats contain all fields (previously only the last field was present in the result)
+- Facet pivot stats return `NAN` for a mean value that's NaN
+- Facet pivot stats return an associative array for percentiles
+
+### Changed
+- Facet pivot stats results use the field name instead of `'stats_fields'` as array key
 
 
 ## [6.3.3]